Rename spinner component
