+++ src/components/animations/SpinnerAnimation.vue
